requirements.txt
# Core Dependencies
pyyaml>=6.0
numpy>=1.24.0
sounddevice>=0.4.6
soundfile>=0.12.1

# Voice Activity Detection
webrtcvad>=2.0.10

# Wake Word Detection
pvporcupine>=3.0.0

# Speech Recognition (Whisper)
# Note: whisper.cpp is used via binary, not Python package

# Text-to-Speech (Piper)
# Note: Piper is used via binary, not Python package

# LLM Integration
anthropic>=0.40.0  # For Claude API

# Voice Recognition
speechbrain>=0.5.16
torchaudio>=2.0.0

# Vision (Phase 4)
opencv-python>=4.8.0
face-recognition>=1.3.0
pytesseract>=0.3.10

# Display Control
screeninfo>=0.8.1
python-xlib>=0.33

# Web Search (Phase 4)
requests>=2.31.0
beautifulsoup4>=4.12.0

# Utilities
python-dotenv>=1.0.0
keyring>=24.0.0
gitpython>=3.1.0

# Development
pytest>=7.4.0
black>=23.0.0
